windows-mcp-server/pyproject.toml
[project]
name = "windows-mcp-server"
version = "1.0.0"
description = "A fully functional Windows MCP (Model Context Protocol) server for AI agent integration"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "AGI Agent Automation", email = "dev@agiagentautomation.com" }
]
keywords = ["mcp", "windows", "ai", "automation", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "pywin32>=306; sys_platform == 'win32'",
    "psutil>=5.9.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "types-aiofiles>=23.2.0",
    "types-psutil>=5.9.0",
]

[project.scripts]
windows-mcp-server = "windows_mcp_server.main:main"

[project.urls]
Homepage = "https://github.com/agiagentautomation/windows-mcp-server"
Documentation = "https://github.com/agiagentautomation/windows-mcp-server#readme"
Repository = "https://github.com/agiagentautomation/windows-mcp-server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/windows_mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=windows_mcp_server --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
]
ignore = ["E501"]  # line too long handled by formatter

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]  # allow assert in tests

[tool.black]
line-length = 100
target-version = ["py311"]
